Disable StartupNotify on the desktop entry
Some desktops like KDE show visual feedback for starting applications
(like bouncing the application icon next to the cursor). Tell them that
this hint is not desired as it will appear in the taken screenshot.

This might not be the best way to fix this, but seems to work for now.